use the first token in the file as the module link
When a file imports another file, with syntax like
  import * as x from 'some/path';
we want 'some/path' to refer to a VName that just means "the entire
file".  It doesn't refer to any text in particular, so we just mark
the first letter in the file as the anchor for this.

Previously we used the first 'export' statement in the file, but this
meant that the xref would jump you to a random spot within the file.